Implement active workout session, exercise reordering, and enhanced home stats
- Create `ActiveWorkoutScreen` and `ActiveWorkoutViewModel` to manage session timers, set logging (weight/reps), and workout completion logic.
- Update `GymDatabase` to version 6, introducing `WorkoutHistoryLog` entity for set-level tracking and adding `Converters` for `LocalDate`.
- Implement `saveWorkoutSession` in `WorkoutRepository` to persist workout history along with detailed execution logs.
- Add `getWorkoutDetailItems` and `updateExerciseOrder` to `WorkoutDAO` to support retrieving and reordering exercises within a workout.
- Refactor `WorkoutDetailScreen` to display a reorderable list of exercises and integrate `WorkoutDetailViewModel` with swap logic.
- Update `HomeScreen` and `HomeViewModel` to calculate and display lifetime volume stats using "Fun Weight" comparisons (e.g., Elephant, Tank) in `HomeStatsCard`.
- Enhance `populateDatabase` to seed rich exercise data (including `isGameEnabled` flags) and pre-defined workouts (Full Body, Leg Day).
- Define `ActiveWorkout` route in `Screen` and update `MainScreen` navigation to support starting workouts from Home and Detail views.
- Update `ExerciseEntity` and `WorkoutHistoryEntity` schemas to include gamification parameters and precise volume/duration tracking.

+++ app/src/main/java/com/example/gymapp002/util/FunWeightConverter.kt
@@ -31,7 +31,7 @@ object FunWeightConverter {
      * Toplam ağırlığı (Volume) alır, en uygun nesneye çevirir.
      * Örn: 900 kg kaldırdıysa, 850 kg (Tofaş) sonucunu döner.
      */
-    fun convert(totalVolumeKg: Int): FunWeightResult {
+    fun convert(totalVolumeKg: Double): FunWeightResult {
         // Verilen kilodan küçük veya eşit olan en büyük eşik değerini bul
         // (Floor Entry mantığı)
         val match = milestones.filterKeys { it <= totalVolumeKg }
